+++ fhir_auth/lib/dstu2/smart_client/smart_mobile_client.dart
@@ -1,9 +1,8 @@
 import 'dart:convert';
-// import 'dart:math';
 
 import 'package:collection/collection.dart';
-import 'package:fhir/dstu2.dart';
-import 'package:http/http.dart';
+import 'package:fhir/r4.dart';
+import 'package:http/http.dart' as http;
 import 'package:oauth2_client/oauth2_client.dart';
 import 'package:oauth2_client/oauth2_helper.dart';
 
@@ -55,7 +54,7 @@ class SmartMobileClient extends SmartClient {
 
   @override
   Future<void> initialize() async {
-    await _getEndpoints;
+    await _getEndpoints();
     if (redirectUri != null) {
       client = OAuth2Client(
         /// Just one slash, required by Google specs
@@ -78,31 +77,33 @@ class SmartMobileClient extends SmartClient {
   }
 
   @override
-  Future<Response?> get(String url, {Map<String, String>? headers}) async =>
-      helper?.get(url, headers: headers);
+  Future<http.Response?> get(String url,
+          {Map<String, String>? headers}) async =>
+      await helper?.get(url, headers: headers);
 
   @override
-  Future<Response?> put(String url,
+  Future<http.Response?> put(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.put(url, headers: headers, body: body);
+      await helper?.put(url, headers: headers, body: body);
 
   @override
-  Future<Response?> post(String url,
+  Future<http.Response?> post(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.post(url, headers: headers, body: body);
+      await helper?.post(url, headers: headers, body: body);
 
   @override
-  Future<Response?> delete(String url, {Map<String, String>? headers}) async =>
-      helper?.delete(url, headers: headers);
+  Future<http.Response?> delete(String url,
+          {Map<String, String>? headers}) async =>
+      await helper?.delete(url, headers: headers);
 
   @override
-  Future<Response?> patch(String url,
+  Future<http.Response?> patch(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.patch(url, headers: headers, body: body);
+      await helper?.patch(url, headers: headers, body: body);
 
   /// Request for the CapabilityStatement (or Conformance) and then identifying
   /// the authUrl endpoint & tokenurl endpoing
-  Future<void> get _getEndpoints async {
+  Future<void> _getEndpoints() async {
     if (authUrl != null && tokenUrl != null) {
       return;
     }