+++ fhir_auth/lib/r4/smart_client/smart_mobile_client.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-// import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:fhir/r4.dart';
@@ -55,7 +54,7 @@ class SmartMobileClient extends SmartClient {
 
   @override
   Future<void> initialize() async {
-    await _getEndpoints;
+    await _getEndpoints();
     if (redirectUri != null) {
       client = OAuth2Client(
         /// Just one slash, required by Google specs
@@ -80,31 +79,31 @@ class SmartMobileClient extends SmartClient {
   @override
   Future<http.Response?> get(String url,
           {Map<String, String>? headers}) async =>
-      helper?.get(url, headers: headers);
+      await helper?.get(url, headers: headers);
 
   @override
   Future<http.Response?> put(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.put(url, headers: headers, body: body);
+      await helper?.put(url, headers: headers, body: body);
 
   @override
   Future<http.Response?> post(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.post(url, headers: headers, body: body);
+      await helper?.post(url, headers: headers, body: body);
 
   @override
   Future<http.Response?> delete(String url,
           {Map<String, String>? headers}) async =>
-      helper?.delete(url, headers: headers);
+      await helper?.delete(url, headers: headers);
 
   @override
   Future<http.Response?> patch(String url,
           {Map<String, String>? headers, dynamic body}) async =>
-      helper?.patch(url, headers: headers, body: body);
+      await helper?.patch(url, headers: headers, body: body);
 
   /// Request for the CapabilityStatement (or Conformance) and then identifying
   /// the authUrl endpoint & tokenurl endpoing
-  Future<void> get _getEndpoints async {
+  Future<void> _getEndpoints() async {
     if (authUrl != null && tokenUrl != null) {
       return;
     }