+++ fhir_auth/lib/r4/smart_client/smart_web_client.dart
@@ -83,7 +83,7 @@ class SmartWebClient extends SmartClient {
               .split('&')[0];
           print('AuthCode: $authorizationCode');
           if (authorizationCode != null) {
-            _popupWin.close();
+            _popupWin?.close();
             if (tokenUrl!.isValid) {
               final response = await http.post(
                 tokenUrl!.value!,